+++ src/components/spartan/SComboboxBlock/SComboboxBlock.vue
@@ -1,34 +1,40 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import { SLabel } from '../SLabel';
 import { SCombobox, type TComboboxProps } from '../SCombobox';
-import { HelpAndErrorTexts } from '@internal';
+import { BlockWrapper, type TBlockWrapperProps } from '@internal';
 
-const props = withDefaults(
-    defineProps<
-        Partial<{ errorText: string; helpText: string; label: string }> & Omit<Partial<TComboboxProps>, 'error'>
-    >(),
-    {
-        errorText: undefined,
-        helpText: undefined,
-        label: undefined,
-    },
-);
+defineEmits<{ (event: 'update:modelValue', value: string | number | object): void }>();
+
+const props = defineProps<Partial<TBlockWrapperProps> & Partial<TComboboxProps>>();
+
+const blockWrapperProps = computed(() => ({
+    label: props.label,
+    id: props.id,
+    errorText: props.errorText,
+    helpText: props.helpText,
+}));
 
 const comboboxProps = computed(() => ({
     ...props,
-    error: Boolean(props.errorText),
-    helpText: undefined,
+    error: props.errorText ? Boolean(props.errorText) : props.error,
     label: undefined,
+    helpText: undefined,
+    errorText: undefined,
 }));
 </script>
 
 <template>
-    <div>
-        <SLabel v-if="label" :for="id">{{ label }}</SLabel>
-        <SCombobox v-bind="comboboxProps">
+    <BlockWrapper wrapper="SComboboxBlock" v-bind="blockWrapperProps">
+        <SCombobox
+            class="w-full"
+            v-bind="comboboxProps"
+            @update:model-value="(newValue) => $emit('update:modelValue', newValue)"
+        >
             <slot />
+            <!-- <template #label>
+                <SLabel v-if="label" :for="id">{{ label }}</SLabel>
+            </template> -->
+            <template #button><slot name="button" /></template>
         </SCombobox>
-        <HelpAndErrorTexts :error="errorText" :help="helpText" />
-    </div>
+    </BlockWrapper>
 </template>